mysite/requirements.txt
apache-libcloud==1.0.0
apparmor==2.10.95
apparmor.click==0.3.13.1
apturl==0.5.2
arrow==0.7.0
beautifulsoup4==4.4.1
binaryornot==0.4.0
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
click==6.6
command-not-found==0.3
cookiecutter==1.4.0
cryptography==1.2.3
defer==1.0.6
defusedxml==0.4.1
Django==1.9.6
django-allauth==0.25.2
django-appconf==1.0.2
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.4.1
django-model-utils==2.5
django-registration==2.1
django-userena==2.0.1
django-users2==0.1.16
djangorestframework==3.3.3
docopt==0.6.2
easy-thumbnails==2.3
feedparser==5.1.3
future==0.15.2
guacamole==0.9.2
html2text==2014.12.29
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
jinja2-time==0.1.0
Keras==1.0.4
language-selector==0.1
LibAppArmor==2.10.95
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.6
MarkupSafe==0.23
numpy==1.11.0
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
pipreqs==0.4.2
plainbox==0.25
poyo==0.4.0
psycopg2==2.6.1
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.5.3
python-debian==0.1.27
python-gnupg==0.3.8
python-systemd==231
python3-openid==3.0.10
pytz==2014.10
pyxdg==0.25
PyYAML==3.11
Quamash==0.5.3
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==0.6.1
scikit-learn==0.17.1
scipy==0.17.1
sessioninstaller==0.0.0
six==1.10.0
sklearn==0.0
SQLAlchemy==1.0.9
sqlparse==0.1.18
system-image==3.1
system-service==0.3
Theano==0.8.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.2
whichcraft==0.4.0
xdiagnose==3.8.4
xkit==0.0.0
XlsxWriter==0.7.3
yarg==0.1.9
